Changes to systemd q3a.service.
Use an absolute path as required by systemd. Distro packages will need
to change it. (Reported by @andreyv and @zeroxfourc.)
Link to server documentation on the wiki instead of maintaining a
subset in the service file.

This doesn't affect Q3lite, but added anyway to keep repo's in sync.

ioquake/ioq3@ad10e66