sdl_snd.c should initialize SDL audio without checking SDL_WasInit().
In SDL2, the initialized subsystems are referenced counted, so it's
safe to initialize them twice, and it makes the SDL_QuitSubSystem
during our shutdown correctly decrement the count. Before (as a
probably-harmless bug), it would not increment the refcount if the
subsystem was already initialized, causing problems when it
decremented it later.

ioquake/ioq3@62acfb3