Fix logging out after quitting the game on Raspbian Lite.
Older versions of SDL would leave the tty console in an unusuable
state after exiting the game with Raspbian Lite, so the q3lite
wrapper script would restart the console, logging the user out.
The current version of SDL has fixed the issue, so the command to
restart the tty console can be removed.

Fixes #7

This issue was reported by clodsman